requirements/base.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-unfold==0.41.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==8.13.50
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-environ==0.4.54
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
